sdk/python/requirements.txt
typer>=0.12.3
python-dateutil>=2.9.0

